remove default constructor and ddestructor

+++ Student.cpp
@@ -1,9 +1,9 @@
 #include "Student.hpp"
 
-Student::Student(std::string n, std::string s, uint64_t i):
+Student::Student(std::string n, std::string s, int i):
     Record(n, s), index(i) {}
 
-uint64_t Student::getIndex() const { return index; }
+int Student::getIndex() const { return index; }
 
 void Student::showAll()
 {
